Restore original home component
