+++ src/app/home.old/home.component.html
@@ -1,0 +1,3 @@
+<h1 *ngIf="loggedIn">You're Logged In!</h1>
+<pre *ngIf="loggedIn">{{userInfo}}</pre>
+<button (click)="login()">Login</button>
