+++ src/app/home.old/home.component.tns.html
@@ -1,0 +1,5 @@
+<!--The content below is only a placeholder and can be replaced.-->
+<ActionBar [title]="title"></ActionBar>
+<StackLayout  class="p-20">
+    <Label text="Welcome to {{ title }}!" class="h1 text-center" textWrap="true"></Label>
+</StackLayout>
